Fixed imports, added missing references
